+++ lib/universign/safeguard.rb
@@ -1,4 +1,4 @@
-module ESign
+module Universign
   module Safeguard
     def self.included(klass)
       klass.extend ClassMethods
@@ -16,27 +16,27 @@ module ESign
           raise ex
         end
 
-        known_exception = ESign::ERROR_CODE[ex.faultCode]
+        known_exception = Universign::ERROR_CODE[ex.faultCode]
 
         if known_exception
           raise known_exception
         elsif ex.faultString.include?('Error on document download for this URL')
           url = ex.faultString.match(/<(.+)>/)[1] rescue 'unknown URL'
-          raise ESign::Document::DocumentURLInvalid.new(url)
+          raise Universign::Document::DocumentURLInvalid.new(url)
         elsif ex.faultString.include?('Invalid document URL')
           url = ex.faultString.match(/<(.+)>/)[1] rescue 'unknown URL'
-          raise ESign::Document::DocumentURLInvalid.new(url)
+          raise Universign::Document::DocumentURLInvalid.new(url)
         elsif ex.faultString.include?('Not enough tokens')
-          raise ESign::NotEnoughTokens
+          raise Universign::NotEnoughTokens
         elsif ex.faultString.include?('ID is unknown')
-          raise ESign::Document::UnknownDocument
+          raise Universign::Document::UnknownDocument
         else
           handle_exception(ex, callback)
         end
 
       rescue RuntimeError => ex
         if ex.message.include?('Authorization failed')
-          raise ESign::Client::InvalidCredentials
+          raise Universign::Client::InvalidCredentials
         end
         raise ex
       end