+++ spec/e_sign_spec.rb
@@ -1,7 +1,0 @@
-require 'spec_helper'
-
-describe ESign do
-  it 'has a version number' do
-    expect(ESign::VERSION).not_to be nil
-  end
-end